build on smartos

+++ src/video.h
@@ -26,7 +26,7 @@ typedef int (*frame_iter_t)(video_frame_t *, void *);
 video_t *video_open(const char *);
 int video_iter_frames(video_t *, frame_iter_t, void *);
 double video_framerate(video_t *);
-int64_t video_nframes(video_t *);
+int video_nframes(video_t *);
 void video_free(video_t *);
 
 #endif
